interpretation-engine-sidekick/requirements.txt
docker-py==1.8.1
redis==2.10.5
requests==2.10.0
six==1.10.0
websocket-client==0.37.0